requirements.txt
# python 3.8
--extra-index-url https://download.pytorch.org/whl/cu116
mxnet==1.6.0
torch
torchvision
scikit_learn
scipy
pyyaml
numpy==1.23.1
Pillow